enabled dest rules if istio integration is enabled
